cpp/2019/intcode: change `n_params` return type: `size_t` -> `int`
The `size_t` was polluting stuff with warnings here and there. I am
sure this change can break this code on some platforms, but I don't
really care right now.

+++ cpp/year2019/intcode/intcode.h
@@ -83,7 +83,7 @@ namespace intcode {
   bool position_mode(int instruction, int n);
 
   // Return the number of parameters for the given opcode.
-  size_t n_params(int opcode);
+  int n_params(int opcode);
 
   // Parse a string into a memory.
   memory parse(const std::string& s);
